Rework FileSystemDirectoryReader definition
Previously, FileSystemDirectoryReader IDL type was defined as having
an associated "directory reader" concept type, but (a) the properties
of that concept type were erroneously defined globally and (b) no
indirection from "this" to it's associated concept object to the
properties was done. Simplify by removing the concept object entirely,
and defining the properties as associated with the IDL type object,
properly scoped.

The association of an entry and directory was very loose and done
without naming the property. Explicitly name those.

No normative/observable behavior changes, just a (significant)
clarification in the text.